Misc. reservation display improvements
