+++ lang/lt/entities.php
@@ -29,8 +29,8 @@ return [
     'reservation' => [
         'model' => '{1} rezervacija|[2,9] rezervacijos|[10,*] rezervacijų',
         'managers' => '{1} rezervacijos valdytojas|[2,9] rezervacijos valdytojai|[10,*] rezervacijos valdytojų',
-        'start_time' => 'pradžia',
-        'end_time' => 'pabaiga',
+        'start_time' => 'rez. pradžia',
+        'end_time' => 'rez. pabaiga',
         'resources' => 'rezervuoti ištekliai',
         'is_reservable' => 'ar rezervuojamas?',
         'period' => 'rezervacijos laikotarpis',
